refactor: organise logic_nodes JS logic into folders
