+++ nodes/web/helpers/logic_nodes/main.js
@@ -1,0 +1,4 @@
+import { COLOR_THEMES } from "./style.js";
+import { setDefaults, setNodeColors } from "./utils.js";
+
+export { COLOR_THEMES, setDefaults, setNodeColors };
