+++ nodes/web/helpers/logic_nodes/style.js
@@ -1,0 +1,13 @@
+const COLOR_THEMES = {
+  red: { nodeColor: "#332222", nodeBgColor: "#553333" },
+  green: { nodeColor: "#223322", nodeBgColor: "#335533" },
+  blue: { nodeColor: "#222233", nodeBgColor: "#333355" },
+  pale_blue: { nodeColor: "#2a363b", nodeBgColor: "#3f5159" },
+  cyan: { nodeColor: "#223333", nodeBgColor: "#335555" },
+  purple: { nodeColor: "#332233", nodeBgColor: "#553355" },
+  yellow: { nodeColor: "#443322", nodeBgColor: "#665533" },
+  orange: { nodeColor: "#663322", nodeBgColor: "#995533" },
+  none: { nodeColor: null, nodeBgColor: null }, // no color
+};
+
+export { COLOR_THEMES };
