+++ nodes/web/logic_nodes.js
@@ -1,66 +1,5 @@
 import { app } from "../../../scripts/app.js";
-
-function setDefaults(node) {
-  node.flowInput = undefined;
-  node.flowOutput = undefined;
-  node.end_loop = undefined;
-
-  for (let i = 0; i < node.inputs.length; i++) {
-    const input = node.inputs[i];
-    if (input.name === "flow") {
-      node.flowInput = node.inputs[i];
-    }
-    if (input.name === "end_loop") {
-      node.end_loop = node.inputs[i];
-    }
-  }
-  for (let i = 0; i < node.outputs.length; i++) {
-    const input = node.outputs[i];
-    if (input.name === "flow") {
-      node.flowOutput = node.outputs[i];
-    }
-  }
-  if (node.flowInput !== undefined) {
-    node.inputs = [node.flowInput];
-  } else {
-    node.inputs = [];
-  }
-
-  if (node.flowOutput !== undefined) {
-    node.outputs = [node.flowOutput];
-  } else {
-    node.outputs = [];
-  }
-
-  if (node.end_loop !== undefined) {
-    node.inputs.push(node.end_loop);
-  }
-  node.addInput("init_value_0", "*");
-  node.addOutput("value_0", "*");
-}
-
-const COLOR_THEMES = {
-  red: { nodeColor: "#332222", nodeBgColor: "#553333" },
-  green: { nodeColor: "#223322", nodeBgColor: "#335533" },
-  blue: { nodeColor: "#222233", nodeBgColor: "#333355" },
-  pale_blue: { nodeColor: "#2a363b", nodeBgColor: "#3f5159" },
-  cyan: { nodeColor: "#223333", nodeBgColor: "#335555" },
-  purple: { nodeColor: "#332233", nodeBgColor: "#553355" },
-  yellow: { nodeColor: "#443322", nodeBgColor: "#665533" },
-  orange: { nodeColor: "#663322", nodeBgColor: "#995533" },
-  none: { nodeColor: null, nodeBgColor: null }, // no color
-};
-
-function setNodeColors(node, theme) {
-  if (!theme) {
-    return;
-  }
-  node.shape = "box";
-  if (theme.nodeColor && theme.nodeBgColor) {
-    node.color = theme.nodeColor;
-    node.bgcolor = theme.nodeBgColor;
-  }
-}
+import { COLOR_THEMES, setDefaults, setNodeColors } from "./helpers/logic_nodes/main.js";
 
 const ext = {
   // Unique name for the extension
